Unmap DMA buffer at PCIDevice destruction (#880)
### Issue
N/A

### Description
Lack of munmap for the DMA buffer kept the device file descriptor alive
longer than intended. During unit tests, this had the effect of
preventing resource cleanup between tests. This manifested as unexpected
test failures on WH systems with IOMMU enabled and default (limited)
IOVA range.

### List of the changes
* Use named constants for DMA buffer size
* Explicitly unmap the DMA buffer if it exists
* Remove an unused method on the PCIDevice interface

### Testing
CI and Manual

### API Changes
This PR has API changes: `PCIDevice::get_fd()` is removed. I grepped for
this in tt-metal and tt-exalens projects; they do not appear to be using
it.