Fix blackhole NOC1 translation bugs (#823)
### Issue

#776 

### Description

FIx readout of NOC1 node id for Blackhole on new 80.18 firmware.
Telemetry reader and arc messenger were relying on predefined
coordinates for ARC core, which may not always be true based on noc
translation enabled and/or usage of NOC1. Additionally, mapping of ETH
coordinates in NOC1 pre translated space is not the same as for NOC0
pre-translated space. Disable the readouts for NOC node id test for
NOC1.

### List of the changes

- Dynamic choice of ARC core coordinates based on noc translation
enabled and noc1 usage
- get_noc_translation_enabled for TTDevice
- Reenable NOC1 node id test
- Disable part for harvested ETH cores readouts

### Testing

CI

### API Changes
/