Constrained Cluster Topology (#886)
### Issue
Working on #870 

### Description
So when we set specific target devices, we'd want the returned cluster
topology to match what was open.
This makes getting info much easier for our clients.
Somebody can still open a ClusterDescriptor with a full topology,
outside of Cluster.

### List of the changes
- Implement create_constrained_cluster_descriptor which copies the right
stuff
- Wrote a couple of template filters to help with this.
- Wrote test for it
- Added a new cluster example configuration for easier testing.
- Change Cluster's constructor to constrain its topology if target
devices are provided.
- Cluster won't take ownership of ClusterDescriptor anymore, because it
only makes sense now.
- Minor change in the way opened devices are printed

### Testing
CI tests.
But will also test this as part of:
https://github.com/tenstorrent/tt-metal/pull/22078
And will manually try to run something with a subset of devices

### API Changes
This PR has API changes:
- [ ] (If breaking change) tt_metal approved PR pointing to this branch:
link
- [ ] (If breaking change) tt_debuda approved PR pointing to this
branch: link