Implement remote TTDevice (#887)
### Issue

Initial work towards #730 

### Description

This PR introduces RemoteTTDevice class. The class is derived from base
TTDevice class, in order to match the interface of the class that we use
in order to get all the required information. It is not going to have
PCIDevice and maybe some other things that WH and BH TTDevice has, so
maybe we should think about some other hieararchy in the future. I think
this should be fine for now.

Only basic read/write from/to device functions are implemented as part
of the PR, other function will come in separate PRs.

### List of the changes

- Add RemoteTTDevice class
- Override read/write functions from TTDevice 
- Add a test for read/write

### Testing

CI + additional test

### API Changes
/