Don't constrain ClusterDescriptor if full set of chips (#904)
### Issue
After #870 , and while trying to merge it to main on metal:
https://github.com/tenstorrent/tt-metal/pull/22078

### Description
There is likely a test issue in t3k fabric tests, but I can't find it.
The previous code would tumble the order in unordered_maps, leading to
some failure in said test. I couldn't find the exact issue.

### List of the changes
- If full set of chips, call default copy constructor instead of
function for constraining.

### Testing
I've manually ran this test on a t3k machine, and saw it passing only
after this change.
./build/test/tt_metal/tt_fabric/fabric_unit_tests
--gtest_filter="*Fabric2DPushFixture.TestUnicastRaw"

### API Changes
There are no API changes in this PR.