Bump tt-logger (#923)
### Description
Latest version of `tt-logger` is much simplified code, with compile time
argument checking, and easy support for completely compiling out
irrelevant log levels.

Part of the simplification is dropping support for `log_debug()` and
`log_debug("something")`. It is now required to always specify what
logType you want a message logged as.

### List of the changes
- Bump the version of `tt-logger`
- Anywhere `logType` was not provided, now provide it
- This actually required a few functional fixes (log_fatal was
previously intended to THROW, so swap to TT_THROW)

### Testing
CI

### API Changes
There are no API changes in this PR.