Add more board information to cluster descriptor (#931)
### Issue

#872 

### Description

Add more board information to cluster descriptor. Working towards
constrained topology discovery, having board information encoded to
cluster descriptor is going to be useful. This PR adds board ids, as
well as chip to board mapping to cluster descriptor. Previously, it was
difficult to know which chips are on the same board so that was making
creating constrained topology from logical IDs hard.

### List of the changes

- Add board info to cluster desc
- Add chip to board id to cluster desc
- Add example cluster descriptor
- Add parsing of new information to cluster descriptor

### Testing
CI

### API Changes
/