Implement constrained topology v2 (#937)
### Issue

#870 

### Description

V2 of the work on constrained topology. As mentioned in the previous PR
#932, UMD was doing few unnecessary reads from devices that shouldn't be
included inside topology so it could discard it based on board id.
Partial, but enough to distinguish boards, information about remote
board id is available on local ETH core. Utilize that information to
avoid reads to remote chip.

### List of the changes

- Add reading remote board id from local ETH core
- Use this information to filter topology
- Remove reads from remote chip if not necessary

### Testing
CI. Manually tested on T3K

### API Changes
/