Support SysmemBuffer with unaligned address and size (#968)
### Issue

#860 

### Description

Implement support for SysmemBuffer class when original virtual address
is not aligned. Since KMD mappings need to be page aligned and the size
needs to be multiple of page size, UMD needs to implement the logic of
calling KMD in a proper way even when user passes in something that does
not follow these requirements. However, SysmemBuffer exposes everything
to the user as it's sized with VA and buffer size that user passed in.

### List of the changes

- Add function to align the address
- Add function to resize the mapping if needed
- Make all functions behave like original VA and buffer size is used
- Add tests to showcase this
- Add documentation to SysmemBuffer functions

### Testing
CI. Tested manually on system with IOMMU enabled

### API Changes
/