Fix active eth links for filtering by pci_target_devices (#1015)
### Issue
The issue is in the future which we're trying to avoid

### Description
We had a mismatch in functionality on logical id and pci id filtering,
in how eth links which are outside of filtered cluster are reported.
This is important, since wrong reporting will lead to metal putting its
own eth dispatch firmware on all idle cores, and then without reset you
can't make the whole cluster work again.

### List of the changes
- Change the way we track active_eth_links in topology discovery

### Testing
Tested manually on an IRD T3K machine with -p filter.

### API Changes
There are no API changes in this PR.