Refactor topology discovery class for legacy FW Wormhole (#1023)
### Issue

Working towards #1020 

### Description

Refactor topology discovery class. The class still works for Wormhole
configurations other than 6U. This PR refactors main functions of
TopologyDiscovery class to be as agnostic as possible to arch, version
of FW and other things. Comments are added in .h file which functions
should be overrided and which should be kept same for all
configurations. It also introduces unique ID for all chips that are used
to discover the topology instead of legacy ETH coordinates. ETH
coordinates are used just for tunneling now, that should be deprecated
with fabric.

### List of the changes

- Refactor topology discovery class
- Add functions that should be overrided for different configurations,
main functions should stay the same
- Add comments with explanations and TODOs for newly added functions
- Rely on unique ID of the chip for topology discovery - picked up from
ETH core
- Trim unused fields in EthAddresses

### Testing

Testing manually on wormhole configurations

### API Changes
/