Add test to verify ETH connections (#1032)
### Issue

Need to verify that ETH connections in all cluster descriptors we make
are bidirectional

### Description

Add a test that verifies ETH connections in cluster descriptor all
bidirectional. I made small mistake while refactoring topology which
caused this to not be the case and this test would be helpful in that
case.

### List of the changes

- Add a test that verifies all ETH connections are bidirectional in
cluster desc

### Testing
CI

### API Changes
/