Print UMD perf in markdown style (#1081)
### Issue

#847 

### Description

Add UMD perf print format. Idea is that UMD perf results are printed in
the format such that results can be pasted from logs directly to README
files and form a table for the results. This PR also introduces common
code for UMD benchmarks where some utility that is useful for all the
benchmarks is going to be.

### List of the changes

- Add microbenchmark utils functions
- Print results in desired format in PCIe DMA benchmark
- Print results in desired format in TLB benchmark
- Print results in desired format in IOMMU benchmark

### Testing
manual testing

### API Changes
/