Add UMD perf workflow (#1079)
### Issue

#1052 

### Description

Add Github workflow for UMD perf tests. We are not going to run this on
CI for now, because it takes lot more time than regular CI. Idea is that
someone who wants to see UMD perf can run Github action and just
copy-paste perf if needed from the log to the README table. For now we
are running perf tests just on N150 with IOMMU enabled.

### List of the changes

- Add yaml for build and run benchmarks workflow
- Add yaml for run benchmarks workflow
- Reduce TLB tests in terms of sizes since they take too long. We should
think about proper number of iterations/sizes for these tests

### Testing

Manual testing by running Github actions

### API Changes
/