Update close device to check if dev in long idle and close remote devices first (#1116)
### Issue
https://github.com/tenstorrent/tt-metal/issues/25377#issuecomment

### Description
There was a change to ensure Metal objects are destructed on exit. This
behaviour was causing a hang in tt::Cluster descriptor after pytest has
exception because it invokes `tt-smi reset` first. Seems like the issue
is due to remote/local device accesses when the device was already put
in `LONG_IDLE`

### List of the changes
Update close_device to first close remote chips because they need
corresponding pcie chip to be active and then only try setting power
state and sending risc reset val if the chip (or corresponding pcie
chip) is not in `LONG_IDLE`

### Testing
Used the repro in the Metal issue

### API Changes
(When making API changes, don't merge this PR until tt_metal and
tt_debuda PRs are approved.)
(Then merge this PR, change the client PRs to point to UMD main, and
then merge them.)
(Remove this line if untrue) There are no API changes in this PR.
(Remove following lines if untrue) This PR has API changes:
- [ ] (If breaking change) tt_metal approved PR pointing to this branch:
link
- [ ] (If breaking change) tt_debuda approved PR pointing to this
branch: link