Implement default TopologyDiscovery behaviour (#1113)
### Issue

#1076 

### Description

Implement default behaviour for TopologyDiscovery. Default behaviour
should be that UMD sees only the cards connected to the host via PCIe,
except in TG (4U galaxy) case. This is done in a way that if we see a TG
board we always accept it, otherwise we just use cards connected to host
via PCIe.

### List of the changes

- Add offsets to read board id high from ETH core. This represents board
type on non 6U wormhole configs
- Integrate this info into checking whether board id should be accepted
- Make checks for TG and other configs separately

### Testing

CI + manual testing on all configs we have

### API Changes
/