Add proper waiting for ETH to TTDevice (#1126)
### Issue

Add proper wait for ETH training to TTDevice

### Description

Aside from waiting for ETH heartbeat, add waiting for ETH cores to get
out of unknown state. Apart from that, if training on 6U fails, don't
report the error since idle ETH cores are not used in metal on 6U, and
failed training might be the case we hit for multi user case.

### List of the changes

- Move wait for ETH training to TTDevice
- Skip throwing the error if training fails for 6U

### Testing
CI

### API Changes
/