Add debug logs to CI runs (#1161)
### Issue
No specific issue, general debuggability. Related to #1002 

### Description
There are two switches which enable debugging. One is at compile time,
and currently it is turned on by default on Debug configuration, and
turned off by default for other configurations. The second switch is the
one which I'm introducing in this PR on our CI runs. The effect should
be as follows:
For our regular Release CI runs, no effect.
If Debug run is requested through CI, we should see additional logs,
helping with the debuggability of issues directly from the CI.

### List of the changes
- Add log-level option for run-tests.yaml and build_and_run_tests. It
can be set now when scheduling a workflow.
- Set it to info as default
- Set TT_LOGGER_LEVEL to corresponding log level
- Increase timeout to 60 if log level is higher, so that we have a
chance at seeing something

### Testing
Scheduled several runs on CI to confirm the behavior:
- Release info (normal run):
https://github.com/tenstorrent/tt-umd/actions/runs/16770564380
- Release trace (should still be like normal run):
https://github.com/tenstorrent/tt-umd/actions/runs/16770568725
- Debug debug (shows debug logs) (with increased timeout of 60):
https://github.com/tenstorrent/tt-umd/actions/runs/16779797819
- Debug trace (shows trace logs):
https://github.com/tenstorrent/tt-umd/actions/runs/16770579234

### API Changes
There are no API changes in this PR.