Move remote transfer cores to RemoteCommunication (#1148)
### Issue
Related to https://github.com/tenstorrent/tt-umd/issues/995

### Description
Eth cores to use for remote communication are ideologically part of
RemoteCommunication class, not LocalChip.
This is on a path for removing dependancy from RemoteWormholeTTDevice to
LocalChip.

### List of the changes
- Move remote eth cores related functions to RemoteCommunication
- Now RemoteChip has to implement set_remote_ethernet_cores since it has
to update its RemoteCommunication class object
- LocalChip also still has to do this, since LocalChip holds
RemoteCommunication object for broadcast purposes.
- Callsites have to be changed, since now this has to be called for
remoteChip (or remote communication) and not localchip
- RemoteWormholeTTDevice has to accept RemoteCommunication object, it
can't construct it on it's own. The RemoteCommunication has to have its
remote transfer eth cores already set before it is passed to the
TTDevice since the constructor uses RemoteCommunication (Note: this
might be changed with some of Nenad's changes in
https://github.com/tenstorrent/tt-umd/issues/1162, but we can change
this later again, not a problem).

### Testing
CI tests

### API Changes
This PR has API changes but they aren't directly used by tt-metal. It
does affect how tt-smi will use these.