Enable ASan and TSan builds in our CI (#638)
### Issue
#627

### Description
As a follow up from #628 and chatting with Andrew, the decision is to
run ASan on Pull requests, since they should be faster but may uncover
many bugs. Pushes on main branch will run TSan in addition to ASan. I've
also added a workflow_dispatch to be able to run any combination of
build types quickly.

### List of the changes
- Added build-type to artefact name so that multiple build types can run
as part of the same workflow.
- minor changes in build-tests.yml
- Minor changes in run-tests.yml, added build-type so it knows which
artefact to download, and removed workflow-dispatch option from it since
it doesn't make sense.
- Changed build-and-run-all-tests not to be scheduled on pr and push,
but to be able to be called from another workflow. Timeout should also
be passed to it.
- Added build-and-run-all-tests-multiple-builds.yml which can run
multiple build-and-run-all-tests with specific builds. It runs ASan in
case of PR, ASan, TSan in case of push to main, and any combination when
manually ran.

### Testing
On PR test, which should run only ASan build can be seen on this PR's CI
For pushing to main, we'll have to merge this PR and see what happens.

Following is a workflow dispatch test. It was run on "Check runners"
since you have to merge a yaml with workflow_dispatch to main, for it to
become available for a manual run even on the development branch.
Therefore I just practically highjacked another .yaml temporarily just
for testing. This workflow dispatch test dispatched all builds:
https://github.com/tenstorrent/tt-umd/actions/runs/14214500087
new run: https://github.com/tenstorrent/tt-umd/actions/runs/16498645891

### API Changes
There are no API changes in this PR.