Some coordinate modifications (#1173)
### Issue
On a path of #687 and a bit of #797 

### Description
Further modification of cleaning up coord internaly.
I've chosen for TLBManager to use tt_xy_pair instead of CoreCoord,
although both options make sense to me.
Also cleaned up some other stuff along the way

### List of the changes
- TLBManager accepts translated coords always, not virtual coords for
its internal map
- Use the same translation function in Chip class as other
CoreCoord->tt_xy_pair transformations are using
- Some code, mainly in tests, used translation where it wasn't needed
anymore, so cleaned that up a bit

### Testing
Existing CI tests should be fine.

### API Changes
There are slight TLBManager API changes in this PR, but this is not used
in clients.