Temp changes to make clients build (#1218)
### Issue
Related to 

### Description
This is a follow up change which makes clients build after all the
changes made in this PR stack:
https://github.com/tenstorrent/tt-umd/pull/1193
https://github.com/tenstorrent/tt-umd/pull/1194
https://github.com/tenstorrent/tt-umd/pull/1195
https://github.com/tenstorrent/tt-umd/pull/1196
https://github.com/tenstorrent/tt-umd/pull/1197
https://github.com/tenstorrent/tt-umd/pull/1198
https://github.com/tenstorrent/tt-umd/pull/1217
https://github.com/tenstorrent/tt-umd/pull/1219

I separated this change as a separate one, so that git diff properly
shows renames (otherwise it would show this change as a modification to
original files + new files, which are actually just renames).
The motivation for this change is to allow two phase API change, this
won't break the builds, the once metal switches to new usages, we can
remove all these intermediate artifacts.

### List of the changes
- Add many header files which just point to another header file, so that
the build doesn't break.
- Update CMakeLists.txt accordingly.

### Testing
Code builds

### API Changes
This PR has API changes which make clients build with the new structure
of header files.