Release automation (#1227)
### Issue
https://github.com/tenstorrent/tt-umd/issues/1178
https://github.com/tenstorrent/tt-umd/issues/1226

### Description
Release automation workflow, which builds all artifacts, and spawns all
content for the release.
The automation also sends a message to slack. And it is triggered when
VERSION file is changed on main branch.
Note that .deb and manylinux wheel packages will come up in later PRs

### List of the changes
- Introduced VERSION file which holds raw version
- Consume version in CMakeLists and pyproject.toml
- CHANGELOG changed to affect new releases
- Added automation workflow which does a few things
  - Builds artefacts (cpp build archives and python wheel)
- Extracts a version, extracts changelog from the CHANGELOG file. Gets
info from git commits for more detailed changelog info
  - Creates a draft release (which should be manually confirmed)
  - Sends a slack message with release info
- The workflow will run on all pushes to main branch which change
VERSION file

### Testing
Here's an example run:
https://github.com/tenstorrent/tt-umd/actions/runs/17249365568
Example release:
https://github.com/tenstorrent/tt-umd/releases/tag/untagged-d06f28a6a8e0a6b4e57c
Example slack msg:
https://tenstorrent.slack.com/archives/C07L16Z59AP/p1756195134111059 -
although it was updated since this was posted, you can find the slack
message in the workflow job, it is printed at some point

### API Changes
There are no API changes in this PR.