Fix silicon system leaking into simulation runs (#1257)
### Issue
Fixes #1238 

### Description
The logic was recreated. The simulation/mock path now doesn't depend on
the system. The original idea was that it might be good to create a set
of mock chips which match what is visible on the system. However, this
proven to be very unintuitive.

### List of the changes
- Rewrite logic in the constructor.
- Most importantly, removed the branch that if no custom cluster
descriptor was passed on a machine with silicon, and simulation chip was
requested, the umd would use system generated cluster descriptor

### Testing
CI testing.
Manual test that simulation still works from tt-metal

### API Changes
This PR has behavior changes for the Cluster api which is explained
above. Not changes required in clients.