Integrate JTAG I/O into Cluster. (#1248)
### Description
Jtag has recently been added as a possible communication protocol in
tt_device. This PR raises it to Cluster so users are able to create
Cluster that works with either PCIe or JTAG.

Currently, because of lack of testing resources, remote communication
with JTAG could not be tested, therefore it wasn't implemented.

This isn't the final version, yet just a step towards the main goal of
making cluster completely independent from PCIe if that is ever needed
in UMD.

Changed Jtag to keep a library handle as a singleton.

### List of the changes
Cluster,  LocalChip, TopologyDiscovery, TopologyDiscoveryWormhole,
- Added communication device type in constructors and performed topology
discovery/creation of chip/cluster based on the communication type.
  
ClusterDescriptor - Added IODeviceType as a field so cluster can fetch
it in its methods to branch different initializations/actions based on
io device type.
ClusterOptions - Added IODeviceType and Jtag_target_devices so future
implementations could potentially use both pci and jtag on a cluster.
Jtag -> Updated Jtag to hold the library handle as a singleton. 
JtagDevice -> Updated JtagDevice to hold the currently opened device id
as a static variable shared by threads in a single process to avoid
accidental device closing by other components in cluster that use the
same shared library.

### Testing
Added a test in test_jtag that checks the basic operations with JTAG
cluster on a single Wormhole n150, and n300.