Add max AICLK value and 18.8 firmware support (#1244)
### Issue

Support getting maximum possible AICLK and together with that add class
for supporting fw properly, working towards #1233

### Description

This PR adds the support for getting maximum possible AICLK on the
device. Since that is released in 18.8 firmware, there are additional
classes that needed to be introduced in order to make everything work
with other fw versions as well. So the PR renames classes for FW support
such that the name of the class represents the latest FW version it
supports. Reading out maximum AICLK is put into FirmwareInfoProvider

### List of the changes

- Get maximum AICLK from telemetry
- Make TTDevice call into FirmwareInfoProvider for it
- Rename Wormhole legacy FW to Wormhole 18.3
- Add Wormhole 18.7 firmware info provider

### Testing
CI

### API Changes
/

---------

Co-authored-by: Bojan Roško <156314064+broskoTT@users.noreply.github.com>