TTSim CI (#1247)
### Issue
#1277 

### Description
Establish a CI for ttsim. Build and run commands are fairly simple.
Note that the ttsim ref is hardcoded in tt_metal workflow, we extract
that ref, so that new changes on ttsim don't unintentionally break the
build.

### List of the changes
- Add ttsim build client workflow
- Runs only wh for now, bh is not passing.

### Testing
The added CI passes

### API Changes
There are no API changes in this PR.