Populate ethernet_connections_to_remote_devices with Logical Eth Channels for P150 (#1297)
### Issue
No Ticket.

### Description
- `ethernet_connections_to_remote_devices` in `topology.cpp` is
currently populated with the physical Ethernet Channel ID for the remote
device
- This is okay for WH systems, since there is no ethernet harvesting. On
BH systems, with harvested ethernet, this causes issues, since users
expect UMD to provide logical ethernet channel IDs for all connection
endpoints
- For neighbor chips on the local host,
`TopologyDiscovery::patch_eth_connections()` ensures that logical
channels are reported
- This cannot be done for chips on the remote host, as the ethernet
harvesting mask is unavailable
- Add an API to query the remote logical ethernet channel, specifically
for P150

### List of the changes
- Add `TopologyDiscovery::get_logical_remote_eth_channel` API
- On WH this is a passthrough to
`TopologyDiscovery::get_remote_eth_channel`
- On BH this API reads the address shared in the Eth Mem Map shared here
https://tenstorrent.atlassian.net/browse/SYS-2064
- Specifically on BH, we need to add 4 (this is specifically for P150)
to the value that is read back, since the "logical" space captured at
this address and the logical IDs that we expect are shifted
- This currently only works with P150, which is not a real blocker
since we are not using P100 or P300 for Scale-Out. An assert has been
added to reflect this