Avoid creating socket when using libttsim.so path (#1309)
### Issue
https://github.com/tenstorrent/ttsim/issues/36

### Description
Socket was being created even when it was unused. This also led to
creation of a large number of unused pthreads inside the nng library and
extra confusing debug output at startup time, and likely slowed down
startup time somewhat as well.

### List of the changes
Split most of the SimulationHost init out of the constructor so that it
can be called conditionally.

### Testing
Ran programming examples locally on simulator

### API Changes
/