Fix new soft reset api for simulation/ttsim (#1346)
### Issue
Unblocking PR https://github.com/tenstorrent/tt-metal/pull/27737

### Description
The code was to specific on the requested RiscType. So when I changed
::ALL_TENSIX to ::ALL to be less confusing in tt_metal PR, this lead to
this branch not being supported by sim.
I've now implemented the logic such that only for Quasar and for DM core
reset one command is called, in any other case the old one is called.
This is more flexible (also more error prone, but I don't think that
will be the problem at this layer).
Also added a ttsim branch for this.

### List of the changes
- Invert the ifelses, such that now dm core reset is the first case
checked
- In all other cases (compared to just being specific on ::ALL_TENSIX or
::BRISC) call the old function for assert/deassert
- Implement ARCH formatter, I saw that actually the runtime code fails
on throwing the exception because this wasn't implemented

### Testing
tt_metal run including ttsim:
https://github.com/tenstorrent/tt-metal/actions/runs/17867710209

### API Changes
This PR has API changes for simulator for newly introduced
assert/deassert api.