Fixing the return of grid size (#1363)
### Issue
[Semaphores out of
bound](https://github.com/tenstorrent/tt-metal/issues/22411)

### Description
When requesting a grid size of cores that are not arranged in a grid
(e.g. ETH) the function throws a null dereferencing exception instead of
returning a value. Both tt_metal and ttnn agree that returning a grid
size (0,0) is more appropriate.

### List of the changes
Both grid_size_map and harvested_grid_size_map now map (0,0) as grid
size for the cores that are not arranged in a grid

### Testing