Separate RTL sim and TT sim paths (#1364)
### Issue
#1271 
#822 

### Description
These two implementations have a lot of if/elses, making the separation
explicit with different implementation classes.
This will make it easier to make subsequent changes which should affect
only one implementation going forward.

### List of the changes
- Rename SimulationDevice to SimulationChip
- Introduce RTLSimulationChip and TTSimulationChip which are derived
classes, with each of the branches copied to these classes

### Testing
Existing CI tests

### API Changes
Changed the API for simulation and how it is used.