Add caching of TLBs in Chip class (#1426)
### Issue

Splitting #959 

### Description

Add caching of TLBs inside Chip class. This is going to be used when we
fully integrate TLBs from KMD. Chip will need to cache TLBs in order to
not always map and unmap the TLBs from KMD. Chip will map one WC TLB for
memory access (L1, DRAM etc...) and one UC TLB for register access.

### List of the changes

- Add field for cached TLB WC and UC
- Add method for getting cached TLB

### Testing
This was tested as part of #959 

### API Changes
/