Move soc descriptor serialize test out of baremetal (#1444)
### Issue

#1369 

### Description

Move soc descriptor serialize test to Cluster API test group. This was
previously in baremetal tests which wasn't right place for it, since it
creates Cluster object. I was not able to reproduce the error in the
issue on P100, it might be that it got fixed by some other commit.

### List of the changes

- Move test to test_cluster.cpp

### Testing
CI

### API Changes
/