Improve TopologyDiscovery with getting ID for chips (#1478)
### Issue

#1464 

### Description

Even if I feel that this issue is not a UMD problem, rather galaxy not
being flashed with proper FW, there is room for improvement for getting
ID of the chips that are not connected to any other chips. So far we
took board ID and that can be wrong when all chips are disconnected.

### List of the changes

- Add function for getting ID of the chip that is not connected to other
chips
- Take asic ID for blackhole
- Still take board ID for wormhole

### Testing
CI

### API Changes
/