Fix multiple artifacts when running tests (#1482)
### Issue
Broken after bumping the VERSION

### Description
The problem is with a runner which doesn't cleanup folders from previous
runs.
So the runner will end up with multiple tt-umd wheel versions after
bumping the VERSION.
This is only an issue with CIv1 runner, which we currently use for P300.

### List of the changes
- Remove wheel folder before building a new wheel, just in case
- Remove download folder just before downloading a new artifact, which
should fix this issue.

### Testing
CI passes on this PR

### API Changes
There are no API changes in this PR.