Add verification for harvesting info of Tenstorrent cards (#1487)
### Issue

We have some cards in the fleet that don't follow the spec for number of
available Tensix, DRAM or ETH cores. Even if UMD can work with all of
those cards, upper layers of stack have some expectations on card
configuration so they might see diff in perf or something else and be
surprised.

### Description

Check number of harvested units for Tensix, DRAM and ETH. Every card has
the spec on TT website on number of harvested units expected. Check
these units and log warning if something is not as expected. We don't
want to block people to work on these cards but it would be nice to let
user know the card is not matching official product spec.

### List of the changes

- Split verify functions in cluster descriptor into multiple functions
- Add function to check expected number of harvested tensix
- Add function to check expecetd number of harvested DRAM
- Add function to check expected number of harvested ETH
- Use these functions to verify harvesting info in cluster descriptor

### Testing

CI + local testing

### API Changes
/