Remove unused target_chip (#1521)
### Issue


### Description
This is already embedded inside remote communication and is not needed
in remote chip as well.
This is probably a leftover from when we switched to using TTDevice
instead of Chip for remote communication.

### List of the changes
- Removed this argument from Remote TTDevice constructors, and where it
was used

### Testing
None, code builds

### API Changes
This PR has API changes, but nobody uses this api directly