Add tt_kmd_lib files (#1556)
### Issue

Work towards #1199 

### Description

Add files implemented by @joelsmithTT, known as tt kmd lib. These files
were implemented in https://github.com/tenstorrent/tt-kmd/pull/101. The
discussions led to the idea to integrate this into UMD for now, as first
tester of these files. This is going to build up on work done in #959.
This PR is meant just to include these files, there will be a follow up
PR that integrates usage of these functions to UMD.

### List of the changes

- Add tt_kmd_lib.c 
- Add tt_kmb_lib.h
- Fix cmake files to be able to build these files inside UMD

### Testing
/

### API Changes
/